+++ test/unit/mail_2_rmq/unit_test_run.sh
@@ -5,11 +5,9 @@
 #   is located at.
 
 test/unit/mail_2_rmq/archive_email.py
-test/unit/mail_2_rmq/camelize.py
 test/unit/mail_2_rmq/check_nonprocess.py
 test/unit/mail_2_rmq/connect_process.py
 test/unit/mail_2_rmq/filter_subject.py
-test/unit/mail_2_rmq/get_email_addr.py
 test/unit/mail_2_rmq/get_text.py
 test/unit/mail_2_rmq/help_message.py
 test/unit/mail_2_rmq/load_cfg.py